+++ next-cloudinary/src/components/CldImage/CldImage.tsx
@@ -1,8 +1,8 @@
-import React, { useState } from 'react';
+import React, { useState, useCallback } from 'react';
 import Image, { ImageProps } from 'next/image';
 import { getTransformations } from '@cloudinary-util/util';
 import { transformationPlugins } from '@cloudinary-util/url-loader';
-import type { ImageOptions } from '@cloudinary-util/url-loader';
+import type { ImageOptions, ConfigOptions } from '@cloudinary-util/url-loader';
 
 import { pollForProcessingImage } from '../../lib/cloudinary';
 
@@ -11,6 +11,7 @@ import { cloudinaryLoader } from '../../loaders/cloudinary-loader';
 export type CldImageProps = Omit<ImageProps, 'src'> & ImageOptions & {
   src: string;
   preserveTransformations?: boolean;
+  config?: ConfigOptions;
 };
 
 const CldImage = (props: CldImageProps) => {
@@ -26,20 +27,19 @@ const CldImage = (props: CldImageProps) => {
       }
       CLD_OPTIONS.push(prop);
     });
-  })
+  });
 
   // Construct the base Image component props by filtering out Cloudinary-specific props
 
   const imageProps = {
     alt: props.alt,
-    src: props.src
+    src: props.src,
   };
 
   (Object.keys(props) as Array<keyof typeof props>)
     .filter(key => !CLD_OPTIONS.includes(key))
     // @ts-expect-error
     .forEach(key => imageProps[key] = props[key]);
-
 
   const defaultImgKey = (Object.keys(imageProps) as Array<keyof typeof imageProps>).map(key => `${key}:${imageProps[key]}`).join(';');
   const [imgKey, setImgKey] = useState(defaultImgKey);
@@ -62,7 +62,7 @@ const CldImage = (props: CldImageProps) => {
 
   if (props.preserveTransformations) {
     try {
-      const transformations = getTransformations(props.src);
+      const transformations = getTransformations(props.src).map(t => t.join(','));
       // @ts-expect-error
       cldOptions.rawTransformations = [...transformations.flat(), ...(props.rawTransformations || [])];
     } catch(e) {
@@ -74,22 +74,49 @@ const CldImage = (props: CldImageProps) => {
    * handleOnError
    */
 
-  interface HandleOnError {
-    target: any;
-  }
+  async function onError(options: React.SyntheticEvent<HTMLImageElement, Event>) {
+    let pollForImage = true;
 
-  async function handleOnError(options: HandleOnError) {
-    const result = await pollForProcessingImage({ src: options.target.src })
+    if ( typeof props.onError === 'function' ) {
+      const onErrorResult = props.onError(options);
+
+      if ( typeof onErrorResult === 'boolean' && onErrorResult === false ) {
+        pollForImage = false;
+      }
+    } else if ( typeof props.onError === 'boolean' && props.onError === false ) {
+      pollForImage = false;
+    }
+
+    // Give an escape hatch in case the user wants to handle the error themselves
+    // or if they want to disable polling for the image
+
+    if ( pollForImage === false ) return;
+
+    const image = options.target as HTMLImageElement
+    const result = await pollForProcessingImage({ src: image.src })
+
     if ( result ) {
       setImgKey(`${defaultImgKey};${Date.now()}`);
     }
   }
 
+  const handleOnError = useCallback(onError, [pollForProcessingImage, defaultImgKey]);
+
+  // Copypasta from https://github.com/prismicio/prismic-next/pull/79/files
+  // Thanks Angelo!
+  // TODO: Remove once https://github.com/vercel/next.js/issues/52216 is resolved.
+
+  let ResolvedImage = Image;
+
+  if ("default" in ResolvedImage) {
+    ResolvedImage = (ResolvedImage as unknown as { default: typeof Image }).default;
+  }
+
   return (
-    <Image
+    <ResolvedImage
       key={imgKey}
       {...imageProps}
-      loader={(loaderOptions) => cloudinaryLoader({ loaderOptions, imageProps, cldOptions })}
+      loader={(loaderOptions) => cloudinaryLoader({ loaderOptions, imageProps, cldOptions, cldConfig: props.config })}
       onError={handleOnError}
     />
   );