+++ next-cloudinary/src/types/player.ts
@@ -13,6 +13,7 @@ export interface CloudinaryVideoPlayerOptions {
   muted?: boolean;
   publicId: string;
   secure?: boolean;
+  transformation?: Array<object> | object;
 }
 
 export interface CloudinaryVideoPlayerOptionsColors {